Move buildSupportedNFT
